Refactor 'lib.pathops' to 'lib.skia'
